moved logic to util file
