+++ apps/frontend/src/components/ViewFormsList/ViewFormsList.tsx
@@ -23,6 +23,7 @@ import ViewFormsOptions from './ViewFormsOptions';
 import { formSchema, addressSchema, AdminNotes } from '../../types/formSchema';
 import { CSVDownload, CSVLink } from 'react-csv';
 import { Asserts } from 'yup';
+import { convertFormsToCSVFormat } from '../../utils/convertToCsv';
 
 export default function ViewFormsList() {
   const [forms, setForms] = useState<FormData[]>([]);
@@ -60,50 +61,6 @@ export default function ViewFormsList() {
   useEffect(() => {
     getForms();
   }, []);
-
-  const convertFormsToCSVFormat = () => {
-    return allForms.map((form) => {
-      return {
-        id: form.id,
-        childsName: form.financialAssistanceForm.childsName,
-        dob: form.financialAssistanceForm.dob,
-        guardianName: form.financialAssistanceForm.guardianName,
-        childsDiagnosis: form.financialAssistanceForm.childsDiagnosis,
-        otherDiagnosis: form.financialAssistanceForm.otherDiagnosis,
-        dateOfDiagnosis: form.financialAssistanceForm.dateOfDiagnosis,
-        childsPhysician: form.financialAssistanceForm.childsPhysician,
-        hospital: form.financialAssistanceForm.hospital,
-        otherHospital: form.financialAssistanceForm.otherHospital,
-        hospitalAddress: convertAddressToString(
-          form.financialAssistanceForm.hospitalAddress
-        ),
-        requestedGrantAmount: `$${form.financialAssistanceForm.requestedGrantAmount}`,
-        intendedUseOfGrant: form.financialAssistanceForm.intendedUseOfGrant,
-        medicalProfessionalPhone:
-          form.financialAssistanceForm.medicalProfessionalPhone,
-        medicalProfessionalEmail:
-          form.financialAssistanceForm.medicalProfessionalEmail,
-        notes: form.financialAssistanceForm.notes,
-        date: form.financialAssistanceForm.date,
-        ...convertAdminNotesToString(form.adminNotes),
-      };
-    });
-  };
-
-  const convertAddressToString = (hospital: Asserts<typeof addressSchema>) => {
-    return `${hospital.street}, ${hospital.city} ${hospital.state}, ${hospital.zipcode}`;
-  };
-
-  const convertAdminNotesToString = (adminNotes: AdminNotes) => {
-    let notes = {};
-    adminNotes.forEach((adminNote, index) => {
-      notes = {
-        ...notes,
-        [`admin_note_${index}`]: `${adminNote.note} (${adminNote.updatedAt})`,
-      };
-    });
-    return notes;
-  };
 
   return (
     <Box p={1}>
@@ -146,7 +103,7 @@ export default function ViewFormsList() {
 
           <Button>
             <CSVLink
-              data={convertFormsToCSVFormat()}
+              data={convertFormsToCSVFormat(allForms)}
               filename="tswgo_all_forms"
               target="_blank"
               style={{
